# out/spring/src/main/java/org/openapitools/api/CameraApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Camera;
import org.openapitools.model.Coordinates;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-02-18T17:28:49.134660478Z[Etc/UTC]")
@Validated
@Tag(name = "camera", description = "Managing camera")
public interface CameraApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /camera
     * Create a new available camera
     *
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "addCamera",
        description = "Create a new available camera",
        tags = { "camera" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/camera"
    )
    
    default ResponseEntity<Void> addCamera(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /camera
     * Returns a list of all cameras
     *
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "getCamera",
        description = "Returns a list of all cameras",
        tags = { "camera" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Camera.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/camera",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Camera>> getCamera(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"name\" : \"camera1\", \"location\" : { \"latitude\" : 41.4, \"longitude\" : 2.1 }, \"id\" : 11 }, { \"name\" : \"camera1\", \"location\" : { \"latitude\" : 41.4, \"longitude\" : 2.1 }, \"id\" : 11 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /camera/{cameraId}
     * Update
     *
     * @param cameraId ID of camera to update (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "updateCamera",
        description = "Update",
        tags = { "camera" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/camera/{cameraId}"
    )
    
    default ResponseEntity<Void> updateCamera(
        @Parameter(name = "cameraId", description = "ID of camera to update", required = true, in = ParameterIn.PATH) @PathVariable("cameraId") Long cameraId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /camera/{cameraId}/location
     * Update
     *
     * @param cameraId ID of camera to update (required)
     * @param coordinates The new location of the camera (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "updateCameraLocation",
        description = "Update",
        tags = { "camera" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/camera/{cameraId}/location",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> updateCameraLocation(
        @Parameter(name = "cameraId", description = "ID of camera to update", required = true, in = ParameterIn.PATH) @PathVariable("cameraId") Long cameraId,
        @Parameter(name = "Coordinates", description = "The new location of the camera", required = true) @Valid @RequestBody Coordinates coordinates
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}